grounding_dino_detection/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.5.0
numpy>=1.20.0
groundingdino-py>=0.1.0